splash screen, ui

+++ lib/pages/cashInflow.dart
@@ -1,6 +1,7 @@
 import 'package:accounts_book/models/acct_class.dart';
 import 'package:accounts_book/models/trxn_class.dart';
 import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
 import 'package:intl/intl.dart';
 
 import '../database.dart';
@@ -28,10 +29,20 @@ class _CashInflowState extends State<CashInflow> {
       trxns=t;
     });
   }
+  void showToast(String toastmsg) {
+    Fluttertoast.showToast(
+        msg: toastmsg,
+        toastLength: Toast.LENGTH_SHORT,
+        gravity: ToastGravity.BOTTOM,
+        timeInSecForIosWeb: 1,
+        backgroundColor: Colors.black54,
+        textColor: Colors.white
+    );
+  }
   gettotal() async{
     var t=(await SQLiteDbProvider.db.gettotal(acct.id, 1))[0]["total"];
     setState(() {
-      total=double.parse(t.toStringAsFixed(2));
+      total=t !=null ? double.parse(t.toStringAsFixed(2)) : 0;
     });
   }
   @override
@@ -98,6 +109,7 @@ class _CashInflowState extends State<CashInflow> {
                                         _sourcecontroller.clear();
                                         _amountcontroller.clear();
                                         setState(() {});
+                                        showToast("Transaction Added");
                                       }
                                       );
                                       Navigator.of(context).pop();
@@ -119,7 +131,7 @@ class _CashInflowState extends State<CashInflow> {
                   future: trxns,
                   builder: (context, AsyncSnapshot snapshot){
                     if(snapshot.hasError) print(snapshot.error);
-                    if(snapshot.hasData){
+                    if(snapshot.hasData && snapshot.data.length>0){
                       return SingleChildScrollView(
                           scrollDirection: Axis.horizontal,
                           child: DataTable(
@@ -206,6 +218,7 @@ class _CashInflowState extends State<CashInflow> {
                                                               _sourcecontroller.clear();
                                                               _amountcontroller.clear();
                                                               setState(() {});
+                                                              showToast("Transaction Updated");
                                                             }
                                                             );
                                                             Navigator.of(context).pop();
@@ -228,11 +241,42 @@ class _CashInflowState extends State<CashInflow> {
                                           ),
                                           ElevatedButton(
                                             onPressed: () {
-                                              SQLiteDbProvider.db.delete(snapshot.data[index].id).then((value) {
-                                                gettrxn();
-                                                gettotal();
-                                                setState(() {});
+                                              showDialog(context: context, builder: (context) {
+                                                return AlertDialog(
+                                                  title: const Text("Confirm Delete"),
+                                                  content: const Text("Do you want to delete this Account"),
+                                                  actions: [
+                                                    ElevatedButton(
+                                                      child: const Text('Yes'),
+                                                      onPressed: () {
+                                                        SQLiteDbProvider.db.delete(snapshot.data[index].id).then((value) {
+                                                          gettotal();
+                                                          gettrxn();
+                                                          setState(() {});
+                                                          showToast("Transaction Deleted");
+                                                        });
+                                                        Navigator.of(context).pop();
+                                                      },
+                                                      style: ButtonStyle(
+                                                        backgroundColor:  MaterialStateProperty.all(Colors.red[300]),
+                                                        overlayColor:  MaterialStateProperty.all(Colors.red[500]),
+
+                                                      ),
+                                                    ),
+                                                    ElevatedButton(
+                                                      child: const Text('NO'),
+                                                      onPressed: () {
+                                                        Navigator.of(context).pop();
+                                                      },
+                                                      style: ButtonStyle(
+                                                        backgroundColor:  MaterialStateProperty.all(Colors.lightGreen[300]),
+                                                        overlayColor:  MaterialStateProperty.all(Colors.lightGreen[500]),
+                                                      ),
+                                                    ),
+                                                  ],
+                                                );
                                               });
+
                                             },
                                             child: Icon(Icons.delete_outline, color: Colors.red),
                                             style: ElevatedButton.styleFrom(
@@ -248,7 +292,12 @@ class _CashInflowState extends State<CashInflow> {
                         );
                     }
                     else{
-                      return const Text("Create a new Account");
+                      return Stack(
+                          alignment: Alignment.center,
+                          children: const [
+                            SizedBox(height: 350.0),
+                            Text("Add a new transaction",style: TextStyle(fontSize: 18.0,fontWeight: FontWeight.w500)),]
+                      );
                     }
                   },
                 ),
@@ -256,13 +305,13 @@ class _CashInflowState extends State<CashInflow> {
             ),
             Container(
               width: double.maxFinite,
-              color: Colors.grey,
+              color: Colors.blueAccent,
               padding: EdgeInsets.all(15.0),
               child: Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  const Text("Total : ",style: TextStyle(fontSize: 20.0,fontWeight: FontWeight.w600),),
-                  Text(double.parse(total.toStringAsFixed(2)).toString(),style: const TextStyle(fontSize: 20.0,fontWeight: FontWeight.w600),)
+                  const Text("Total : ",style: TextStyle(fontSize: 20.0,fontWeight: FontWeight.w600,color: Colors.white),),
+                  Text(double.parse(total.toStringAsFixed(2)).toString(),style: const TextStyle(fontSize: 20.0,fontWeight: FontWeight.w600,color: Colors.white),)
                 ],
               ),
             )
