+++ lib/pages/home.dart
@@ -1,6 +1,8 @@
 import 'package:accounts_book/database.dart';
 import 'package:accounts_book/models/acct_class.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
 
 import 'acctbook.dart';
 
@@ -18,6 +20,16 @@ class _HomeState extends State<Home> {
   getaccts() {
     var t=SQLiteDbProvider.db.getAllAccts();
     accts= t;
+  }
+  void showToast(String toastmsg) {
+    Fluttertoast.showToast(
+        msg: toastmsg,
+        toastLength: Toast.LENGTH_SHORT,
+        gravity: ToastGravity.BOTTOM,
+        timeInSecForIosWeb: 1,
+        backgroundColor: Colors.black54,
+        textColor: Colors.white
+    );
   }
   @override
   void initState() {
@@ -46,7 +58,7 @@ class _HomeState extends State<Home> {
                 future: accts,
                 builder: (context, AsyncSnapshot snapshot){
                   if(snapshot.hasError) print(snapshot.error);
-                  if(snapshot.hasData){
+                  if(snapshot.hasData && snapshot.data.length>0){
                     return ListView.builder(
                           itemCount: snapshot.data.length,
                           scrollDirection: Axis.vertical,
@@ -59,16 +71,43 @@ class _HomeState extends State<Home> {
                                   onTap: () {
                                     Navigator.push(context, MaterialPageRoute(builder: (context) => AccountBook(acct: snapshot.data[index])));
                                   },
-                                  title: Text(snapshot.data[index].acct),
+                                  title: Text(snapshot.data[index].acct,style: const TextStyle(fontWeight: FontWeight.w500, fontSize: 20.0),),
                                   trailing: ElevatedButton(
                                     onPressed: () {
-                                        SQLiteDbProvider.db.deleteAcct(snapshot.data[index].id).then((value) {
-                                          getaccts();
-                                          setState(() {});
-                                        });
-                                        ScaffoldMessenger.of(context).showSnackBar(
-                                          const SnackBar(content: Text('Account Deleted')),
+                                      showDialog(context: context, builder: (context) {
+                                        return AlertDialog(
+                                          title: const Text("Confirm Delete"),
+                                          content: const Text("Do you want to delete this Account"),
+                                          actions: [
+                                            ElevatedButton(
+                                              child: const Text('Yes'),
+                                              onPressed: () {
+                                                SQLiteDbProvider.db.deleteAcct(snapshot.data[index].id).then((value) {
+                                                  getaccts();
+                                                  setState(() {});
+                                                });
+                                                showToast('Account Deleted');
+                                                Navigator.of(context).pop();
+                                              },
+                                              style: ButtonStyle(
+                                                backgroundColor:  MaterialStateProperty.all(Colors.red[300]),
+                                                overlayColor:  MaterialStateProperty.all(Colors.red[500]),
+
+                                              ),
+                                            ),
+                                            ElevatedButton(
+                                              child: const Text('NO'),
+                                              onPressed: () {
+                                                  Navigator.of(context).pop();
+                                              },
+                                              style: ButtonStyle(
+                                                backgroundColor:  MaterialStateProperty.all(Colors.lightGreen[300]),
+                                                overlayColor:  MaterialStateProperty.all(Colors.lightGreen[500]),
+                                              ),
+                                            ),
+                                          ],
                                         );
+                                      });
                                     },
                                     child: Icon(Icons.delete_outline,color: Colors.red,),
                                     style: ButtonStyle(
@@ -85,13 +124,18 @@ class _HomeState extends State<Home> {
                     );
                   }
                   else{
-                    return const Text("Create a new Account");
+                    return Stack(
+                        alignment: Alignment.center,
+                        children: const [
+                          SizedBox(height: 350.0),
+                          Text("Create a new Account",style: TextStyle(fontSize: 18.0,fontWeight: FontWeight.w500)),]
+                    );
                   }
                 },
               ),
             ),
           ),
-          Divider(thickness: 3.0,height: 5.0,color: Colors.blueAccent,),
+          const Divider(thickness: 3.0,height: 5.0,color: Colors.blueAccent,),
           Padding(
             padding: const EdgeInsets.all(10.0),
             child: Container(
@@ -123,16 +167,15 @@ class _HomeState extends State<Home> {
                     onTap: () {
                       // Validate returns true if the form is valid, or false otherwise.
                       if (_formKey.currentState!.validate()) {
-                        SQLiteDbProvider.db.insertAcct(Acct(0,_acctcontroller.text,0,0)).then((value) {
+                        SQLiteDbProvider.db.insertAcct(Acct(0,_acctcontroller.text)).then((value) {
                           getaccts();
                           setState(() {
                             _acctcontroller.clear();
                           });
                         }
                       );
-                        ScaffoldMessenger.of(context).showSnackBar(
-                          const SnackBar(content: Text('Account Created')),
-                        );
+                        showToast('Account Created');
+
                       }
                     },
                     child: Icon(Icons.add,color: Colors.grey[600],size: 35.0,),
